So many changes

+++ src/capengine/viewport.cpp
@@ -47,11 +47,10 @@ void Viewport::setY(int y)
  \return
    \li The new point
 */
-Point Viewport::transform(const Point &point)
+Point Viewport::transform(const Point& point)
 {
-  auto vector = m_transformationMatrix * Vector(point.x, point.y, 0, 1);
-  return {static_cast<int>(vector.getX() + 0.5),
-		  static_cast<int>(vector.getY() + 0.5)};
+    auto vector = m_transformationMatrix * Vector(point.x, point.y, 0.0f, 1.0f);
+    return {vector.getX() + 0.5, vector.getY() + 0.5};
 }
 
 Rect Viewport::transformRect(Rect in_rect) const
